==> 1.1. Heróis.ipynb <==
import pandas as pd

microdados_ = pd.read_pickle('microdados_.pkl')
microdados_geo_mun = pd.read_pickle('microdados_geo_mun.pkl')

microdados_.head()

microdados_.columns

microdados_geo_mun.head()

microdados_geo_mun[['CO_MUNICIPIO_ESC', 'NO_MUNICIPIO_ESC', 'nota_media_geral']].groupby(['CO_MUNICIPIO_ESC', 'NO_MUNICIPIO_ESC']).mean().reset_index().sort_values('nota_media_geral', ascending=False)


grupo_1 = microdados_[(microdados_['TP_LOCALIZACAO_ESC']=='Rural')&
            (microdados_['TP_DEPENDENCIA_ADM_ESC'].isin(['Estadual', 'Municipal']))&
            (microdados_['icp']<=4)].sort_values('nota_media_geral', ascending=False)[['NU_INSCRICAO', 'nota_media_geral', 'icp_comp_01', 'icp_comp_02', 'icp_comp_03']]

grupo_0 = microdados_[(microdados_['TP_LOCALIZACAO_ESC']=='Urbana')&
            (microdados_['TP_DEPENDENCIA_ADM_ESC'].isin(['Federal', 'Privada']))&
            (microdados_['icp']>=14)].sort_values('nota_media_geral', ascending=False)[['NU_INSCRICAO', 'nota_media_geral', 'icp_comp_01', 'icp_comp_02', 'icp_comp_03']]


print(f"Percentual de notas acima de 700 - grupo vulnerável: {len(grupo_1[grupo_1['nota_media_geral']>700])/len(grupo_1)}")
print(f"Percentual de notas acima de 700 - grupo controle: {len(grupo_0[grupo_0['nota_media_geral']>700])/len(grupo_0)}")
      

len(grupo_1)

len(grupo_1[grupo_1['nota_media_geral']>700])

str(grupo_1.iloc[0]['NU_INSCRICAO'])

microdados_[microdados_['NU_INSCRICAO']==210056852546.0][['CO_MUNICIPIO_ESC', 'NO_MUNICIPIO_ESC', 'nota_media_geral', 'Q006', 'TP_LOCALIZACAO_ESC', 'TP_DEPENDENCIA_ADM_ESC']]


set(microdados_['Q006'])

microdados_[microdados_['CO_MUNICIPIO_ESC']==2904209.0]['nota_media_geral'].

grupo_1.boxplot(column=['nota_media_geral'])

grupo_1
