==> 03. Indicador de Dedicação (ID).ipynb <==
import pandas as pd

microdados_ = pd.read_pickle('microdados_.pkl')
habitos_estudo = pd.read_csv('microdados_enem_2022/DADOS/QUEST_HAB_ESTUDO.csv', encoding='latin1', sep=';')


microdados_.head()

microdados_.columns

### 01. Índice de Dedicação (ID)

habitos_estudo = habitos_estudo[['NU_INSCRICAO', 'Q005', 'Q007', 'Q008']]
habitos_estudo = habitos_estudo[(pd.isna(habitos_estudo['Q005'])==False)&(pd.isna(habitos_estudo['Q007'])==False)&(pd.isna(habitos_estudo['Q008'])==False)]
habitos_estudo.head()


len(habitos_estudo)

dicio_valor_letra = {'A':1, 'B':2, 'C':3, 'D':4}
def id_(row):
    id_ = dicio_valor_letra[row['Q005']] + dicio_valor_letra[row['Q007']] + dicio_valor_letra[row['Q008']]
    return id_

habitos_estudo['id'] = habitos_estudo.apply(id_, axis=1)


habitos_estudo.head()

microdados_ = microdados_.merge(habitos_estudo[['NU_INSCRICAO', 'id']], how = 'left', on = 'NU_INSCRICAO')
microdados_['id'] = microdados_['id'].fillna('Sem informação')

microdados_.head()

microdados_id = microdados_[microdados_['id']!='Sem informação'][['NU_INSCRICAO', 'nota_media_objetivas', 'nota_media_geral',
       'valor_referencia_renda_familia', 'num_residentes', 'renda_percapita',
       'classe_renda_percapita', 'icp_comp_01', 'icp_comp_02_parcial',
       'icp_comp_02', 'icp_comp_03_parcial', 'icp_comp_03', 'icp', 'regiao', 'TP_COR_RACA', 'id']]

microdados_id.head()


microdados_id[['id', 'nota_media_geral']].groupby(['id']).mean()

microdados_id[['icp', 'id']].groupby(['icp']).mean()

microdados_id_dif_icp = microdados_id[microdados_id['id']==12][['icp', 'nota_media_geral']].groupby(['icp']).mean().reset_index().merge(
microdados_id[microdados_id['id']==3][['icp', 'nota_media_geral']].groupby(['icp']).mean().reset_index(), how = 'inner', on = 'icp')
microdados_id_dif_icp['dif'] = microdados_id_dif_icp['nota_media_geral_x'] - microdados_id_dif_icp['nota_media_geral_y']
microdados_id_dif_icp['icp'] = microdados_id_dif_icp['icp'].apply(lambda x: 'ICP '+str(x))
microdados_id_dif_icp[['icp', 'dif']]


microdados_.to_pickle('microdados_.pkl')
